changed foreign key type Int to Long

+++ app/src/main/java/com/wagit/desktrack/data/entities/Registry.kt
@@ -14,7 +14,7 @@ import java.time.LocalDateTime
 )
 data class Registry(
     @PrimaryKey(autoGenerate = true) val id: Long,
-    @ColumnInfo(name = "user_id") val userId: Int,
+    @ColumnInfo(name = "user_id") val userId: Long,
     @ColumnInfo(name = "started_at") val startedAt: LocalDateTime,
     @ColumnInfo(name = "ended_at") val endedAt: LocalDateTime?,
 )
